requirements.txt
absl-py==2.0.0
asttokens==2.0.5
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.5.1
decorator==5.1.1
efficientnet==1.1.1
exceptiongroup==1.1.3
executing==0.8.3
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.45.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
gunicorn==21.2.0
h5py==3.10.0
idna==3.4
imageio==2.33.0
importlib-resources==6.1.1
importlib-metadata==4.11.3 
ipykernel==6.9.1
ipython==8.2.0
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jupyter-client==6.1.12
jupyter-core==4.9.2
jupyterlab-widgets==3.0.9
keras==2.15.0
Keras-Applications==1.0.8
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.2
ml-dtypes==0.2.0
nest-asyncio==1.5.5
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==21.3
pandas==2.1.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.20 
protobuf==4.23.4
psutil==5.8.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.11.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2021.3
pywin32==227
pyzmq==22.3.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0 
stack-data==0.2.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.26
tornado==6.1
tqdm==4.66.1
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.5
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
zipp==3.7.0
